+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -17,6 +17,7 @@
 #include "Rule10p7Check.h"
 #include "Rule10p8Check.h"
 #include "Rule12p3Check.h"
+#include "Rule13p4Check.h"
 #include "Rule14p4Check.h"
 #include "Rule15p1Check.h"
 #include "Rule15p2Check.h"
@@ -73,6 +74,8 @@ public:
         "misrac-rule-10p8");
     CheckFactories.registerCheck<Rule12p3Check>(
         "misrac-rule-12p3");
+    CheckFactories.registerCheck<Rule13p4Check>(
+        "misrac-rule-13p4");
     CheckFactories.registerCheck<Rule14p4Check>(
         "misrac-rule-14p4");
     CheckFactories.registerCheck<Rule15p1Check>(